Inverse Opinon Scores (#26)
Improves how bars are displayed when sorting the top users chart by "Pro-Impeachment Score" which is a derived metric. It keeps the opinion score labels and scale the same for consistency (0: pro-impeachment, 1:pro-trump), but reverses the bar sizes and axis labels.

This is better than before where 0s had no bar height, but may still be confusing for some viewers. There is a simpler way to do this that says just calculate the inverse opinion and use that in the charting. Can be accomplished in the next iteration.